Move common tiff handling to utils.cog
